Extract dashboard search to component #531
