+++ src/app/dashboard/components/dashboard-search/dashboard-search.component.html
@@ -1,0 +1,6 @@
+<erz-typeahead
+  [typeaheadService]="studentsRestService"
+  [placeholder]="'dashboard.search.placeholder'"
+  (valueChange)="$event && navigateToDossier($event)"
+>
+</erz-typeahead>
