+++ src/app/dashboard/components/dashboard/dashboard.component.scss
@@ -19,11 +19,6 @@
   flex: 2;
 }
 
-.search-box {
-  padding: $spacer * 0.75;
-  background: $dark-sand;
-}
-
 @include media-breakpoint-up(lg) {
   .columns {
     flex-direction: row;
